Added new Repositories and Service Files. Moved UserServiceImpl to /serviceImpl
