+++ src/main/java/com/sprint/hibernate/service/MarathonService.java
@@ -1,0 +1,4 @@
+package com.sprint.hibernate.service;
+
+public interface MarathonService {
+}
