+++ src/main/java/com/sprint/hibernate/service/TaskService.java
@@ -1,0 +1,4 @@
+package com.sprint.hibernate.service;
+
+public interface TaskService {
+}
